long is 64 bits on mac, int is 32

+++ ui/sdk/include/event.h
@@ -10,7 +10,7 @@ struct Event {
 
 // 对象继承链内部路由的消息
 struct Msg {
-  long message;
+  int message;
 };
 
 }
